+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Adding rewind client API ([#253](https://github.com/microsoft/durabletask-java/pull/253))
+
 ## v1.7.0
 * Add descriptive error when orchestration type is not registered ([#261](https://github.com/microsoft/durabletask-java/pull/261))
 * Update all dependencies to latest versions ([#260](https://github.com/microsoft/durabletask-java/pull/260))